Commenting policy regarding private methods/fields
Changed * **XMLDOC** so that private methods and parameters don't need xmldoc and neither do fields. Of course developers can still add them if they are needed, but this is not needed to be enforced.

Removed from ADD-VALUE  " even if the comment is trivial on trivial code (XML documentation comments)"  because we should not be adding trivial comments to trivial code.  And having a rule like this will make some developers add comments to code that intentionally didn't have it.